requirements.txt
Flask==3.0.0
flask-cors==4.0.0
opencv-python>=4.9.0
numpy>=2.0.0
pandas>=2.2.0
tensorflow>=2.16.0
Pillow>=10.2.0
fer==22.5.0
requests==2.31.0
python-dotenv==1.0.0

